dblist: update nuke engine

+++ META-INF/scripts/bin/nukeengine.sh
@@ -1,5 +1,7 @@
 #! /tmp/busybox sh
 
 cat /tmp/dblist | while read line; do
-    /tmp/busybox rm -rf "/system_root/system/$line"
+    if [ -n "$line" ]; then
+        /tmp/busybox rm -rf "/system_root/system/$line"
+    fi
 done
